Complete rewrite of test/run.sh
Output is more useful (IMO). Everything is silenced for the duration of
the tests but a count of tests/failures and any errors are output at the
end. Pass --verbose to watch stdout during the tests.

Logic is also seperated into core setup/runners and individual test
definitions. Should be more easily extensible.