Facebook logout support and access token.
This patch renames "facebookUrl" to "facebookLogin" and adds
"facebookLogout".  The latter logs out the user from Facebook as
well to comply with Facebook's policies.

As a bonus, we also save the user's access token into their
session and export "getFacebookAccessToken".  Besides being
useful for the application, this is necessary for the new logout
action.

This patch also bumps the version to 0.8 since we have renamed
"facebookUrl".