static: Fix a bug with development mode cache control headers
During development mode, the static subsite was serving Cache-Control
headers for far in the future.  This is what we want to do during
production since we export routes with the etag as a query parameter,
but during development these headers will cause the browser to not
even make a request.  During development, we want the browser to
always make a request and if the content is the same we will respond
with a 304.  This is now the current behavior.