static: fix the build on GHC 7.4
There were two build issues on GHC 7.4: using atomicModifyIORef'
and ByteString.Lazy.toStrict, both of which were missing.  These
are now fixed.

In addition, looking at the IORef code more closely,
we want the quite a bit of strictness in the IORef.
The common case is that the widget content already exists in the map
(every reload of a page will call embedStaticContent), but until we
force the map the thunks holding the duplicated generated content will
be kept around, leaking memory.  This will be the common situation since
the vast majority of the time the content already exists.  Since
the containers module does not have a strict map until 0.5, use
unordered containers which does have a strict map.