Converted link to use Text instead of HttpMethod
HttpMethod limited the available http methods. Since rest isn't the only
standard and the standards can change I removed standard restful http
methods and just used text. Forms only support GET and POST by default,
so if the method isn't one of those it will place it in the URL as a
_method.