Don't hardcode templates directory; make it a setting.
This allows to load yesod applications into ghci together
with other packages, for example for `stack repl`.

Until now, the templates directory was hardcoded, which
required one to be inside the directory containing `templates`
for `stack repl .` to work.
That automatically forbade using `stack repl` in projects
that have more than one Yesod project.

This commit makes it possible to pass in the templates directory,
so that it can be passed as an absolute directory, so that it
would work from anywere.
To do this, the TemplateHaskell `location` function can be used,
for example for the following directory tree:

  myproject.cabal
  templates/
  src/Settings.hs

the file `src/Settings.hs` could contain:

  import Language.Haskell.TH (location, stringE, loc_filename)
  import System.FilePath     (takeDirectory, (</>))

  widgetFileSettings :: WidgetFileSettings
  widgetFileSettings = def
    { wfsTemplatesDirectory = $(location >>= \loc -> stringE $ takeDirectory (loc_filename loc) </> ".." </> "templates")
    }

This commit breaks `yesod`'s API and thus requires a version bump,
because the `globFile` function was exported and has its type
signature changed.