Make site's logger type be instance-based
At the moment, all sites are tied to using the fast-logger `Logger` type
for Yesod's logging functions, including `makeLogger` and
`messageLoggerSource`. This represents an obstacle to logging through
other libraries which are incompatible with the fast-logger API.

This replaces the use of the concrete `Logger` type with an associated
type family on the `Yesod` class which allows sites to use their choice
of logger representation for Yesod's logging functions. The associated
type family is declared in such a way that existing instances should
work with no further effort.

This commit introduces a change in semantics around how logging is
shared between the Yesod site and the underlying WAI server - because
WAI is still tied to fast-logger, it's not possible to have them share
the logging config in general. Instead, the logger passed to WAI is the
default stdout logger.